+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#1.2.10
+- Added button to the issue sidebar for adding the product manager application review comment
+- Added a confirmation button for sending participation comments
+- Updated UI to hide / show participation comments
+
 #1.2.9
 - Updated logic for WAQ groupings